add e2e test for index
